requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
chromadb
google-genai>=0.1.2
langchain-community
langchain
pypdf
